Added new main method and olfModel
The entire model is based off another paper with the following html:
https://academic.oup.com/chemse/article/27/9/789/305840#4002403
It has the complete g protein activation, ACII, cAMP and CNG, and negative feedback. It's different from what we currently have but the input and output of the system is that the input is odorant concentration and time, whilst the output is cAMP concentration, activated CNG channels (related to membrane voltage and so action potentials), active Ca-CL channels, intracellular Ca2+. And a couple other ones we don't care about. Would appreciate if someone could go over the paper and have comments on if the model is sufficient.